change the int to double in reminder

+++ sample-spring-boot-kitchensink/src/main/java/com/example/bot/spring/SQLDatabaseEngine.java
@@ -189,8 +189,8 @@ public class SQLDatabaseEngine extends DatabaseEngine {
 
 
 	//Query the target diet plan info from "diet_plan" table and return
-	ArrayList<Integer> search_plan(String user_id) throws Exception {
-		ArrayList<Integer> plan_info = new ArrayList<Integer>(); // store the query result from plan table
+	ArrayList<Double> search_plan(String user_id) throws Exception {
+		ArrayList<Double> plan_info = new ArrayList<Double>(); // store the query result from plan table
 		try {
 			//connect to the database with table: diet_plan
 			Connection connection = this.getConnection();
@@ -199,9 +199,9 @@ public class SQLDatabaseEngine extends DatabaseEngine {
 			stmt.setString(1, user_id);//1 is the param location/index
 			ResultSet rs = stmt.executeQuery();
 			while (rs.next()) {
-				plan_info.add(rs.getInt(1));//protein
-				plan_info.add(rs.getInt(2));//fat
-				plan_info.add(rs.getInt(3));//sugar
+				plan_info.add(rs.getDouble(1));//protein
+				plan_info.add(rs.getDouble(2));//fat
+				plan_info.add(rs.getDouble(3));//sugar
 			}
 			rs.close();
 			stmt.close();
@@ -217,8 +217,8 @@ public class SQLDatabaseEngine extends DatabaseEngine {
 	}
 
 	//Query the current diet status info from "diet_conclusion" table and return
-	ArrayList<Integer> search_current(String user_id, String date) throws Exception {
-		ArrayList<Integer> current_info = new ArrayList<Integer>(); // store the query result from current table
+	ArrayList<Double> search_current(String user_id, String date) throws Exception {
+		ArrayList<Double> current_info = new ArrayList<Double>(); // store the query result from current table
 		try {
 			//connect
 			Connection connection = this.getConnection();
@@ -227,9 +227,9 @@ public class SQLDatabaseEngine extends DatabaseEngine {
 			stmt.setString(2, date);
 			ResultSet rs = stmt.executeQuery();
 			while (rs.next()) {
-				current_info.add(rs.getInt(1));//protein
-				current_info.add(rs.getInt(2));//fat
-				current_info.add(rs.getInt(3));//sugar
+				current_info.add(rs.getDouble(1));//protein
+				current_info.add(rs.getDouble(2));//fat
+				current_info.add(rs.getDouble(3));//sugar
 			}
 			rs.close();
 			stmt.close();
